requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
pycryptodome>=3.20.0
curl_cffi>=0.6.0
pywin32>=306
